+++ paper/src/main/java/com/badbones69/crazyvouchers/api/objects/Voucher.java
@@ -277,7 +277,8 @@ public class Voucher {
             }
         }
 
-        this.totalWeight = this.randomCommands.stream().filter(filter -> filter.getWeight() >= 0.0D).mapToDouble(VoucherCommand::getWeight).sum();
+        // if a prize has a weight greater than 0.0, include it with the total weight.
+        this.totalWeight = this.randomCommands.stream().filter(filter -> filter.getWeight() > 0.0D).mapToDouble(VoucherCommand::getWeight).sum();
     }
 
     public @NotNull String getArgument(@NotNull final ItemStack item) {
@@ -604,7 +605,8 @@ public class Voucher {
         if (this.randomCommands.isEmpty()) return;
 
         // dispatch commands without a weight option randomly
-        final List<VoucherCommand> randomCommands = this.randomCommands.stream().filter(filter -> filter.getWeight() > 0.0D).toList();
+        // if the prize weight is less than 0.0, grab it.
+        final List<VoucherCommand> randomCommands = this.randomCommands.stream().filter(filter -> filter.getWeight() < 0.0D).toList();
 
         if (!randomCommands.isEmpty()) {
             final VoucherCommand randomCommand = randomCommands.get(Methods.getRandom(randomCommands.size()));
@@ -613,8 +615,8 @@ public class Voucher {
         }
 
         // dispatch commands while accounting for the weight on each one.
-        // if a section has Weight, and a list of commands. all those commands will execute if the Weight is picked.
-        final List<VoucherCommand> chanceCommands = this.randomCommands.stream().filter(filter -> filter.getWeight() <= 0.0D).toList();
+        // if a prize has a weight greater than 0.0, grab it.
+        final List<VoucherCommand> chanceCommands = this.randomCommands.stream().filter(filter -> filter.getWeight() > 0.0D).toList();
 
         if (!chanceCommands.isEmpty()) {
             Methods.dispatch(player, getCommand(chanceCommands).getCommands(), placeholders, true);